requirements.txt
# Complete AI Trading Bot Requirements
# This file contains all necessary dependencies for the refactored bot

# Core Trading Libraries
ccxt>=4.0.0
asyncio-throttle>=1.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Machine Learning
torch>=1.12.0
torchvision>=0.13.0
scikit-learn>=1.1.0
xgboost>=1.6.0
gymnasium>=0.26.0
stable-baselines3>=1.6.0

# Optimization
optuna>=3.0.0

# Monitoring and Metrics
influxdb-client>=1.36.0
psutil>=5.9.0

# Messaging
python-telegram-bot>=20.0.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=1.10.0

# Date/Time Processing
python-dateutil>=2.8.0

# Memory Profiling
memory-profiler>=0.60.0

# Async and Concurrency
aiohttp>=3.8.0
aioredis>=2.0.0

# Database
sqlite3  # Built into Python

# HTTP Requests
requests>=2.28.0
httpx>=0.24.0

# Utilities
tqdm>=4.64.0
click>=8.1.0

# Development and Testing
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=22.10.0
flake8>=5.0.0

# Optional Dependencies (for enhanced functionality)
pyarrow>=10.0.0  # For Parquet support
plotly>=5.10.0   # For visualization
jupyter>=1.0.0   # For analysis notebooks

# Note: Some dependencies may require specific Python versions
# Python 3.8+ is recommended for full compatibility